Move to statically-compiled module
Boost can be a bear of a dependency, as can a C++11 compiler.
This removes the need for a compilation environment on the
end-user system.

Well, this is true if we've pre-built a binary for the particular
end-user system. If not, then it's still hell.
