+++ binding.gyp
@@ -2,9 +2,7 @@
   "targets": [
     {
       "target_name": "<(module_name)",
-      "sources": [
-        "mmap-object.cc"
-      ],
+      "type": "loadable_module",
       "cflags_cc": [
         "-Wall",
         "-Werror",
@@ -12,45 +10,25 @@
         "-fexceptions",
         "-frtti"
       ],
-      "link_settings": {
-        "ldflags": [
-          "-L/usr/lib64/boost148"
-        ]
-      },
-      "conditions": [
-        [
-          "OS=='mac' or '<!(echo $MACHTYPE)'=='x86_64-redhat-linux-gnu'",
-          {
-            "libraries": [
-              "-lboost_system-mt",
-              "-lboost_thread-mt"
-            ]
-          }
-        ],
-        [
-          "OS=='linux' and not '<!(echo $MACHTYPE)'=='x86_64-redhat-linux-gnu'",
-          {
-            "libraries": [
-              "-lboost_system",
-              "-lboost_thread"
-            ]
-          }
-        ]
-      ],
-      "xcode_settings": {
-        "OTHER_CFLAGS": [
-          "-I/usr/local/opt/boost155/include",
-          "-Wno-unused-local-typedefs"
-        ],
-        "OTHER_LDFLAGS": [
-          "-L/usr/local/opt/boost155/lib"
-        ],
-        "GCC_ENABLE_CPP_EXCEPTIONS": "YES",
-        "GCC_ENABLE_CPP_RTTI": "-frtti"
-      },
+      "sources": ["main.cc"],
       "include_dirs": [
-        "<!(node -e \"require('nan')\")",
-        "/usr/include/boost148/"
+        "<!(node -e \"require('nan')\")"
+      ],
+      "dependencies": [ "mmo-lib"]
+    },
+    {
+      "target_name": "mmo-lib",
+      "cflags_cc": [
+        "-Wall",
+        "-Werror",
+        "-O3",
+        "-fexceptions",
+        "-frtti"
+      ],
+      "type": "static_library",
+      "sources": [ "mmap-object.cc"],
+      "include_dirs": [
+        "<!(node -e \"require('nan')\")"
       ]
     },
     {
