+++ mmap-object.h
@@ -1,0 +1,109 @@
+#include <nan.h>
+#include <boost/interprocess/managed_mapped_file.hpp>
+#include <boost/interprocess/managed_shared_memory.hpp>
+#include <boost/interprocess/containers/string.hpp>
+#include <boost/unordered_map.hpp>
+
+#define MINIMUM_FILE_SIZE 500 // Minimum necessary to handle an mmap'd unordered_map on all platforms.
+#define DEFAULT_FILE_SIZE 5ul<<20 // 5 megs
+#define DEFAULT_MAX_SIZE 5000ul<<20 // 5000 megs
+
+namespace bip=boost::interprocess;
+using namespace std;
+
+typedef bip::managed_shared_memory::segment_manager segment_manager_t;
+
+template <typename StorageType> using SharedAllocator =
+  bip::allocator<StorageType, segment_manager_t>;
+
+typedef SharedAllocator<char> char_allocator;
+
+typedef bip::basic_string<char, char_traits<char>, char_allocator> shared_string;
+typedef bip::basic_string<char, char_traits<char>> char_string;
+
+#define UNINITIALIZED 0
+#define STRING_TYPE 1
+#define NUMBER_TYPE 2
+
+class WrongPropertyType: public exception {};
+class FileTooLarge: public exception {};
+
+class Cell {
+private:
+  char cell_type;
+  union values {
+    shared_string string_value;
+    double number_value;
+    values(const char *value, char_allocator allocator): string_value(value, allocator) {}
+    values(const double value): number_value(value) {}
+    values() {}
+    ~values() {}
+  } cell_value;
+  Cell& operator =(const Cell&) = default;
+  Cell(Cell&&) = default;
+  Cell& operator=(Cell&&) & = default;
+public:
+  Cell(const char *value, char_allocator allocator) : cell_type(STRING_TYPE), cell_value(value, allocator) {}
+  Cell(const double value) : cell_type(NUMBER_TYPE), cell_value(value) {}
+  Cell(const Cell &cell);
+  char type() { return cell_type; }
+  const char *c_str();
+  operator string();
+  operator double();
+};
+
+typedef shared_string KeyType;
+typedef Cell ValueType;
+
+typedef SharedAllocator<pair<KeyType, ValueType>> map_allocator;
+
+struct s_equal_to {
+  bool operator()( const char_string& lhs, const shared_string& rhs ) const;
+  bool operator()( const shared_string& lhs, const shared_string& rhs ) const;
+};
+
+class hasher {
+public:
+  size_t operator() (shared_string const& key) const;
+  size_t operator() (char_string const& key) const;
+};
+  
+typedef boost::unordered_map<
+  KeyType,
+  ValueType,
+  hasher,
+  s_equal_to,
+  map_allocator> PropertyHash;
+
+class SharedMap : public Nan::ObjectWrap {
+public:
+  static NAN_MODULE_INIT(Init);
+
+private:
+  string file_name;
+  size_t file_size;
+  size_t max_file_size;
+  bip::managed_mapped_file *map_seg;
+  PropertyHash *property_map;
+  bool readonly;
+  bool closed;
+  void grow(size_t);
+  void setFilename(string);
+  static NAN_METHOD(Create);
+  static NAN_METHOD(Open);
+  static NAN_METHOD(Close);
+  static NAN_METHOD(isClosed);
+  static NAN_METHOD(get_free_memory);
+  static NAN_METHOD(get_size);
+  static NAN_METHOD(bucket_count);
+  static NAN_METHOD(max_bucket_count);
+  static NAN_METHOD(load_factor);
+  static NAN_METHOD(max_load_factor);
+  static NAN_PROPERTY_SETTER(PropSetter);
+  static NAN_PROPERTY_GETTER(PropGetter);
+  static v8::Local<v8::Function> init_methods(v8::Local<v8::FunctionTemplate> f_tpl);
+  static inline Nan::Persistent<v8::Function> & constructor() {
+    static Nan::Persistent<v8::Function> my_constructor;
+    return my_constructor;
+  }
+};
